+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Change Log
 ==========
 
 # Next Release
+Ensure that access to the constructor is available allowing private constructors to be used
 
 # 0.2.0
 Allow use to optionally use default primitive values when missing from JSON