+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 
 # Next Release
 Ensure that access to the constructor is available allowing private constructors to be used
+When `enableDefaultPrimitiveValues` is set, always treat primitives as though they have a default value
 
 # 0.2.0
 Allow use to optionally use default primitive values when missing from JSON