+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Change Log
 ==========
 
 # Next Release
+
+# 0.3.0
 Ensure that access to the constructor is available allowing private constructors to be used
 When `enableDefaultPrimitiveValues` is set, always treat primitives as though they have a default value
 Only fetch the delegate adapter when the property is going to be serialized/deserialized